Add files via upload
The main web applcation was done in Python with the help of ArcGIS API for python. This API provides
interactive display of geographic information through a composition of web layers, basemap and other
similar tools for geographic mapping. We also used JavaScript to implement certain UI elements like popups
and legend for the layers in the application.
The HUC-4 watershed shapefiles for entire case studies uploaded in the ArcGIS online website as a
public layer. Also, the rainfall and runoff stations and its associated dataset uploded in ArcGIS online
website as feature layers and then all information deployed in the web app using APIs.